refactor static runtime memory planner in order to make usable in jit as well
[ghstack-poisoned]
